Improve disk space checking for syncstream, remove for replays
Filesystem::DiskSpace now returns an unsigned long long to correctly report
free disk space above 4GB. This may break compilation on ancient compilers
that don't support long long.

The disk space check is now on the basis of how much data has been written
since the last check, and not on a time basis.

+++ src/io/filesystem/filesystem.h
@@ -117,7 +117,7 @@ struct FileSystem {
     static std::string FS_FilenameWoExt(char const *);
 	static std::string GetHomedir();
 
-	virtual unsigned long DiskSpace() = 0;
+	virtual unsigned long long DiskSpace() = 0;
 
 protected:
 	///To get a filesystem, use the Create methods
